+++ lib/models/api/autotest_api_model.dart
@@ -12,10 +12,9 @@ final class AutotestFullModel {
   final String? id;
   final bool? isFlaky;
 
-  factory AutotestFullModel.fromJson(Map<String, dynamic> json) {
-    return AutotestFullModel(
-        json['id'].toString(), bool.tryParse(json['isFlaky'].toString()));
-  }
+  factory AutotestFullModel.fromJson(Map<String, dynamic> json) =>
+      AutotestFullModel(
+          json['id'].toString(), bool.tryParse(json['isFlaky'].toString()));
 
   const AutotestFullModel(this.id, this.isFlaky);
 }
@@ -24,10 +23,9 @@ final class AutoTestRelatedToTestResult {
   final String? externalId;
   final bool? isDeleted;
 
-  factory AutoTestRelatedToTestResult.fromJson(Map<String, dynamic> json) {
-    return AutoTestRelatedToTestResult(json['externalId'].toString(),
-        bool.tryParse(json['isDeleted'].toString()));
-  }
+  factory AutoTestRelatedToTestResult.fromJson(Map<String, dynamic> json) =>
+      AutoTestRelatedToTestResult(json['externalId'].toString(),
+          bool.tryParse(json['isDeleted'].toString()));
 
   const AutoTestRelatedToTestResult(this.externalId, this.isDeleted);
 }