+++ lib/services/api/test_run_api_service.dart
@@ -5,23 +5,22 @@ import 'dart:io';
 
 import 'package:adapters_flutter/converters/test_result_converter.dart';
 import 'package:adapters_flutter/managers/config_manager.dart';
+import 'package:adapters_flutter/managers/log_manager.dart';
 import 'package:adapters_flutter/models/api/autotest_api_model.dart';
 import 'package:adapters_flutter/models/api/test_run_api_model.dart';
-import 'package:adapters_flutter/models/config/merged_config_model.dart';
+import 'package:adapters_flutter/models/config_model.dart';
 import 'package:adapters_flutter/models/exception_model.dart';
 import 'package:adapters_flutter/models/test_result_model.dart';
 import 'package:http/http.dart';
-import 'package:logger/logger.dart';
 
-final Logger _logger = Logger();
+final _logger = getLogger();
 
-Future<void> createEmptyTestRunAsync(final MergedConfigModel config) async {
+Future<void> createEmptyTestRunAsync(final ConfigModel config) async {
   try {
     final headers = {
       'accept': 'application/json',
       'Content-Type': 'application/json',
-      'Authorization': 'PrivateToken ${config.privateToken}',
-      'host': Uri.tryParse(config.url!)?.host ?? ''
+      'Authorization': 'PrivateToken ${config.privateToken}'
     };
 
     final request =
@@ -36,8 +35,11 @@ Future<void> createEmptyTestRunAsync(final MergedConfigModel config) async {
     final response = await Response.fromStream(streamedResponse);
 
     if (response.statusCode < 200 || response.statusCode > 299) {
-      throw TmsApiException(
-          'Status code: ${response.statusCode}, Reason: ${response.reasonPhrase}');
+      final exception = TmsApiException(
+          'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
+      _logger.i('$exception.');
+
+      return;
     }
 
     final body = jsonDecode(response.body) as Map<String, dynamic>;
@@ -45,20 +47,19 @@ Future<void> createEmptyTestRunAsync(final MergedConfigModel config) async {
 
     await updateTestRunIdAsync(testRunId);
   } catch (exception, stacktrace) {
-    _logger.i('$exception${Platform.lineTerminator}$stacktrace');
+    _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
   }
 }
 
-Future<List<String>> getTestsFromTestRunAsync(
-    final MergedConfigModel config) async {
-  final List<String> testsFromTestRun = [];
+Future<List<String>> getExternalIdsFromTestRunAsync(
+    final ConfigModel config) async {
+  final List<String> externalIds = [];
 
   try {
     final headers = {
       'accept': 'application/json',
       'Content-Type': 'application/json',
-      'Authorization': 'PrivateToken ${config.privateToken}',
-      'host': Uri.tryParse(config.url!)?.host ?? ''
+      'Authorization': 'PrivateToken ${config.privateToken}'
     };
 
     final request = Request(
@@ -71,8 +72,11 @@ Future<List<String>> getTestsFromTestRunAsync(
     final response = await Response.fromStream(streamedResponse);
 
     if (response.statusCode < 200 || response.statusCode > 299) {
-      throw TmsApiException(
-          'Status code: ${response.statusCode}, Reason: ${response.reasonPhrase}');
+      final exception = TmsApiException(
+          'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
+      _logger.i('$exception.');
+
+      return externalIds;
     }
 
     final body = jsonDecode(response.body) as Map<String, dynamic>;
@@ -91,23 +95,22 @@ Future<List<String>> getTestsFromTestRunAsync(
         continue;
       }
 
-      testsFromTestRun.add(autotest.externalId!);
+      externalIds.add(autotest.externalId!);
     }
   } catch (exception, stacktrace) {
-    _logger.i('$exception${Platform.lineTerminator}$stacktrace');
+    _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
   }
 
-  return testsFromTestRun;
+  return externalIds;
 }
 
 Future<void> submitResultToTestRunAsync(
-    final MergedConfigModel config, final TestResultModel testResult) async {
+    final ConfigModel config, final TestResultModel testResult) async {
   try {
     final headers = {
       'accept': 'application/json',
       'Content-Type': 'application/json',
-      'Authorization': 'PrivateToken ${config.privateToken}',
-      'host': Uri.tryParse(config.url!)?.host ?? ''
+      'Authorization': 'PrivateToken ${config.privateToken}'
     };
 
     final request = Request(
@@ -123,10 +126,13 @@ Future<void> submitResultToTestRunAsync(
     final response = await request.send();
 
     if (response.statusCode < 200 || response.statusCode > 299) {
-      throw TmsApiException(
-          'Status code: ${response.statusCode}, Reason: ${response.reasonPhrase}');
+      final exception = TmsApiException(
+          'Status code: ${response.statusCode}, Reason: "${response.reasonPhrase}".');
+      _logger.i('$exception.');
+
+      return;
     }
   } catch (exception, stacktrace) {
-    _logger.i('$exception${Platform.lineTerminator}$stacktrace');
+    _logger.i('$exception${Platform.lineTerminator}$stacktrace.');
   }
 }